refactorization finish

+++ src/Controller/CommentController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\DTO\DeleteCommentDto;
+use Psr\Log\LoggerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\Request;
@@ -21,13 +22,18 @@ class CommentController extends AbstractController
 	/** @var CommentResponseFactory */
 	private $commentFactory;
 
+	/** @var LoggerInterface */
+	private $logger;
+
 	public function __construct(
 	    CommentFacade $commentFacade,
-        CommentResponseFactory $commentFactory
+        CommentResponseFactory $commentFactory,
+        LoggerInterface $logger
 	)
 	{
 	    $this->commentFacade = $commentFacade;
 	    $this->commentFactory = $commentFactory;
+	    $this->logger = $logger;
 	}
 
 	/**
@@ -36,21 +42,28 @@ class CommentController extends AbstractController
     public function createComment(int $post_id, Request $request): JsonResponse
     {
         $content = $request->request->get('content');
+        $csrf_token = $request->request->get('csrf_token');
         if($content !== null) {
-            try {
-                $comment = $this->commentFacade->createComment(
-                    new CreateCommentDto(
-                        $content,
-                        $post_id
-                    )
-                );
+            if($this->isCsrfTokenValid($this->getUser()->getUsername(), $csrf_token)) {
+                try {
+                    $comment = $this->commentFacade->createComment(
+                        new CreateCommentDto(
+                            $content,
+                            $post_id
+                        )
+                    );
 
-                return new JsonResponse($this->commentFactory->getCommentJson($comment), 200);
-            } catch (\Exception $e) {
-                return new JsonResponse(['message' => $e->getMessage()], 500);
+                    return new JsonResponse($this->commentFactory->getCommentJson($comment), 200);
+                } catch (\Exception $e) {
+                    $this->logger->error($e->getMessage());
+                    return new JsonResponse(['message' => 'Server encountered an error while saving your comment.'], 500);
+                }
+            } else {
+                $this->logger->error('CSRF token invalid: ["token" => '.$csrf_token.'], ["username" => '.$this->getUser()->getUsername().']');
+                return new JsonResponse(['message' => 'Error while deleting comment'], 500);
             }
         } else {
-            return new JsonResponse(['message' => 'null_content'], 500);
+            return new JsonResponse(['message' => 'You cannot post an empty comment!'], 500);
         }
     }
 
@@ -60,23 +73,30 @@ class CommentController extends AbstractController
     public function editComment(int $post_id, int $comment_id, Request $request): JsonResponse
     {
         $content = $request->request->get('content');
+        $csrf_token = $request->request->get('csrf_token');
         $comment_editor_username = $request->request->get('current_user');
         if($content !== null && $comment_editor_username !== null) {
-            try {
-                $comment = $this->commentFacade->editComment(
-                    new EditCommentDto(
-                        $content,
-                        $comment_id,
-                        $comment_editor_username
-                    )
-                );
+            if($this->isCsrfTokenValid($comment_editor_username, $csrf_token)) {
+                try {
+                    $comment = $this->commentFacade->editComment(
+                        new EditCommentDto(
+                            $content,
+                            $comment_id,
+                            $comment_editor_username
+                        )
+                    );
 
-                return new JsonResponse($this->commentFactory->getCommentJson($comment), 200);
-            } catch (\Exception $e) {
-                return new JsonResponse(['message' => $e->getMessage()], 500);
+                    return new JsonResponse($this->commentFactory->getCommentJson($comment), 200);
+                } catch (\Exception $e) {
+                    $this->logger->error($e->getMessage());
+                    return new JsonResponse(['message' => 'Server encountered an error while saving your comment.'], 500);
+                }
+            } else {
+                $this->logger->error('CSRF token invalid: ["token" => '.$csrf_token.'], ["username" => '.$comment_editor_username.']');
+                return new JsonResponse(['message' => 'Error while deleting comment'], 500);
             }
         } else {
-            return new JsonResponse(['message' => 'null_content_or_username'], 500);
+            return new JsonResponse(['message' => 'You cannot post an empty comment!'], 500);
         }
     }
 
@@ -86,22 +106,29 @@ class CommentController extends AbstractController
     public function deleteComment(int $post_id, int $comment_id, Request $request): JsonResponse
     {
         $comment_deleter_username = $request->request->get('current_user');
+        $csrf_token = $request->request->get('csrf_token');
 
         if($comment_deleter_username !== null) {
-            try {
-                $this->commentFacade->deleteComment(
-                    new DeleteCommentDto(
-                        $comment_id,
-                        $post_id,
-                        $comment_deleter_username
-                    )
-                );
-                return new JsonResponse(['message' => 'Success'], 200);
-            } catch (\Exception $e) {
-                return new JsonResponse(['message' => $e->getMessage()], 500);
+            if($this->isCsrfTokenValid($comment_deleter_username, $csrf_token)) {
+                try {
+                    $this->commentFacade->deleteComment(
+                        new DeleteCommentDto(
+                            $comment_id,
+                            $post_id,
+                            $comment_deleter_username
+                        )
+                    );
+                    return new JsonResponse(['message' => 'Success'], 200);
+                } catch (\Exception $e) {
+                    $this->logger->error($e->getMessage());
+                    return new JsonResponse(['message' => 'Server encountered an error while deleting the comment.'], 500);
+                }
+            } else {
+                $this->logger->error('CSRF token invalid: ["token" => '.$csrf_token.'], ["username" => '.$comment_deleter_username.']');
+                return new JsonResponse(['message' => 'Error while deleting comment'], 500);
             }
         } else {
-            return new JsonResponse(['message' => 'null_username'], 500);
+            return new JsonResponse(['message' => 'Log in first!'], 500);
         }
     }
 }
